fix(glc): fix race condition in GetGraphByShareLink - update view count in transaction before return
