+++ website/components/cve-detail-modal.tsx
@@ -9,7 +9,7 @@ import React, { useState, useCallback, useRef, useEffect } from 'react';
 import { Button } from "@/components/ui/button";
 import { Badge } from "@/components/ui/badge";
 import BookmarkStar from '@/components/bookmark-star';
-import { generateURN } from '@/lib/utils';
+import { generateURN, getSeverityVariant, getSeverityLabel, getSeverityColorHex, getCVSSScore, getCVEDescription } from '@/lib/utils';
 import type { CVEItem } from "@/lib/types";
 import { useViewLearnMode } from '@/contexts/ViewLearnContext';
 import { ExpandAllIcon, CollapseAllIcon, CopyIcon } from '@/components/icons';
@@ -21,59 +21,6 @@ interface CVEDetailModalProps {
   cve: CVEItem;
   onClose: () => void;
 }
-
-const getSeverityVariant = (score?: number): "default" | "secondary" | "destructive" | "outline" => {
-  if (!score && score !== 0) return "outline";
-  if (score >= 9.0) return "destructive";
-  if (score >= 7.0) return "default";
-  if (score >= 4.0) return "secondary";
-  return "outline";
-};
-
-const getSeverityLabel = (score?: number): string => {
-  if (!score && score !== 0) return "Unknown";
-  if (score >= 9.0) return "Critical";
-  if (score >= 7.0) return "High";
-  if (score >= 4.0) return "Medium";
-  return "Low";
-};
-
-const getSeverityColorHex = (score?: number): string => {
-  if (!score && score !== 0) return '#9CA3AF';
-  if (score >= 9.0) return '#ef4444';
-  if (score >= 7.0) return '#f59e0b';
-  if (score >= 4.0) return '#fbbf24';
-  return '#10b981';
-};
-
-const getCVSSScore = (cve: CVEItem): number | undefined => {
-  const metricLists = [
-    cve.metrics?.cvssMetricV40,
-    cve.metrics?.cvssMetricV31,
-    (cve.metrics as any)?.cvssMetricV30 || (cve.metrics as any)?.cvssMetricV3,
-    cve.metrics?.cvssMetricV2,
-  ];
-
-  let maxScore: number | undefined = undefined;
-  for (const list of metricLists) {
-    if (!Array.isArray(list)) continue;
-    for (const item of list) {
-      const s = item?.cvssData?.baseScore;
-      if (typeof s === 'number') {
-        if (maxScore === undefined || s > maxScore) maxScore = s;
-      }
-    }
-    if (maxScore !== undefined) break;
-  }
-  return maxScore;
-};
-
-const getDescription = (cve: CVEItem): string => {
-  const desc = cve.descriptions?.find(d => d.lang === 'en') || cve.descriptions?.[0];
-  if (!desc) return 'No description available';
-  const limit = 100;
-  return desc.value.length > limit ? desc.value.substring(0, limit) + '...' : desc.value;
-};
 
 function MetricDetail({ m, metricKey, isOpen, onToggle, onCopy, copiedKeys }: { m: any; metricKey: string; isOpen: boolean; onToggle: () => void; onCopy: (text: string, key: string) => void; copiedKeys: Set<string> }) {
   const cvss = m?.cvssData || {};
@@ -269,7 +216,7 @@ export function CVEDetailModal({ cve, onClose }: CVEDetailModalProps) {
               itemId={cve.id}
               itemType="CVE"
               itemTitle={cve.id}
-              itemDescription={getDescription(cve)}
+              itemDescription={getCVEDescription(cve)}
               viewMode={mode}
             />
             <div className={isLearnMode ? 'learn-minimal' : ''}>Source: {cve.sourceIdentifier || (cve as any).source || 'Unknown'}</div>
