+++ website/components/cve-table.tsx
@@ -24,6 +24,7 @@ import type { CVEItem } from "@/lib/types";
 import { createLogger } from "@/lib/logger";
 import { useViewLearnMode } from '@/contexts/ViewLearnContext';
 import { CVEDetailModal } from '@/components/cve-detail-modal';
+import { getSeverityVariant, getSeverityLabel, getSeverityColorHex, getCVSSScore, getCVEDescription } from '@/lib/utils';
 
 const logger = createLogger("cve-table");
 
@@ -37,65 +38,6 @@ interface CVETableProps {
   onPageSizeChange: (pageSize: number) => void;
   searchQuery?: string;
 }
-
-// Helper utilities extracted to module scope so they're not recreated on each render
-const getSeverityVariant = (score?: number): "default" | "secondary" | "destructive" | "outline" => {
-  if (!score && score !== 0) return "outline";
-  if (score >= 9.0) return "destructive";
-  if (score >= 7.0) return "default";
-  if (score >= 4.0) return "secondary";
-  return "outline";
-};
-
-const getSeverityLabel = (score?: number): string => {
-  if (!score && score !== 0) return "Unknown";
-  if (score >= 9.0) return "Critical";
-  if (score >= 7.0) return "High";
-  if (score >= 4.0) return "Medium";
-  return "Low";
-};
-
-// Map severity to a small color for subtle row accent (hex colors to avoid depending on Tailwind palette variations)
-const getSeverityColorHex = (score?: number): string => {
-  if (!score && score !== 0) return '#9CA3AF'; // gray-400
-  if (score >= 9.0) return '#ef4444'; // red-500
-  if (score >= 7.0) return '#f59e0b'; // amber-500 (high)
-  if (score >= 4.0) return '#fbbf24'; // yellow-400 (medium)
-  return '#10b981'; // green-500 (low)
-};
-
-// Prefer newer metrics (4.0 -> 3.1 -> 3.0 -> 2.0) and return the highest baseScore available
-const getCVSSScore = (cve: CVEItem): number | undefined => {
-  const metricLists = [
-    cve.metrics?.cvssMetricV40,
-    cve.metrics?.cvssMetricV31,
-    // some payloads may use cvssMetricV3 or cvssMetricV30
-    (cve.metrics as any)?.cvssMetricV30 || (cve.metrics as any)?.cvssMetricV3,
-    cve.metrics?.cvssMetricV2,
-  ];
-
-  let maxScore: number | undefined = undefined;
-  for (const list of metricLists) {
-    if (!Array.isArray(list)) continue;
-    for (const item of list) {
-      const s = item?.cvssData?.baseScore;
-      if (typeof s === 'number') {
-        if (maxScore === undefined || s > maxScore) maxScore = s;
-      }
-    }
-    // if we already found a score from a higher-preference metric family, prefer it
-    if (maxScore !== undefined) break;
-  }
-  return maxScore;
-};
-
-const getDescription = (cve: CVEItem): string => {
-  const desc = cve.descriptions?.find(d => d.lang === 'en') || cve.descriptions?.[0];
-  if (!desc) return 'No description available';
-  // List view should show a concise summary (shorter for better table fit)
-  const limit = 100;
-  return desc.value.length > limit ? desc.value.substring(0, limit) + '...' : desc.value;
-};
 
 export const CVETable = memo(function CVETable({
   cves,
@@ -137,7 +79,7 @@ export const CVETable = memo(function CVETable({
     return filtered.map((cve) => ({
       id: cve.id,
       score: getCVSSScore(cve),
-      description: getDescription(cve),
+      description: getCVEDescription(cve),
       published: cve.published ? new Date(cve.published).toLocaleDateString() : 'N/A',
       lastModified: cve.lastModified ? new Date(cve.lastModified).toLocaleDateString() : undefined,
       source: cve.sourceIdentifier || (cve as any).source || 'Unknown',
